Update ProcessInvoker to support password input and upgrade version
The main functionality of the ProcessInvoker class was expanded to support password input for executing processes with elevated permissions. This required tweaking of the invoking method, addition of Java libraries, and ensuring appropriate logging. Corresponding changes were also reflected in the WatcherConfiguration and new ProcessInvokerTest classes. The project version was updated to 0.0.2-SNAPSHOT to reflect these changes.